refactor: decouple api from sync
