+++ back-end/app/api/player_routes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify
-from sync.services.player_service import get_player_snapshot
+from coclib.services.player_service import get_player_snapshot
 from coclib.services.loyalty_service import get_player_loyalty
 from ..services.risk_service import get_history, score_breakdown
 from . import API_PREFIX
