+++ coclib/services/__init__.py
@@ -1,0 +1,10 @@
+"""Shared service helpers used by multiple applications."""
+
+from . import player_service, clan_service, war_service, loyalty_service
+
+__all__ = [
+    "player_service",
+    "clan_service",
+    "war_service",
+    "loyalty_service",
+]
