+++ coclib/services/clan_service.py
