+++ coclib/services/player_service.py
@@ -2,7 +2,6 @@ from coclib.utils import safe_to_thread
 from datetime import datetime, timedelta
 import logging
 import os
-import httpx
 from typing import TYPE_CHECKING
 
 from coclib.extensions import db, cache
@@ -13,19 +12,7 @@ from coclib.services.loyalty_service import ensure_membership
 from coclib.utils import normalize_tag
 
 logger = logging.getLogger(__name__)
-SYNC_BASE = os.getenv("SYNC_BASE")
 STALE_AFTER = timedelta(seconds=int(os.getenv("SNAPSHOT_MAX_AGE", "600")))
-
-
-async def _trigger_sync(tag: str) -> None:
-    if not SYNC_BASE:
-        return
-    url = f"{SYNC_BASE.rstrip('/')}/player/{tag}"
-    try:
-        async with httpx.AsyncClient(timeout=10) as client:
-            await client.post(url)
-    except Exception as exc:  # pragma: no cover - best effort
-        logger.warning("Sync request to %s failed: %s", url, exc)
 
 
 async def _fetch_player(tag: str) -> dict:
@@ -178,7 +165,10 @@ async def get_player_snapshot(tag: str) -> "Optional[PlayerDict]":
     row = await safe_to_thread(_latest)
     needs_refresh = row is None or (datetime.utcnow() - row.ts > STALE_AFTER)
     if needs_refresh:
-        await _trigger_sync(norm_tag)
+        try:
+            await get_player(norm_tag)
+        except RuntimeError:
+            pass
         row = await safe_to_thread(_latest)
         if row is None:
             return None
